Ändrade allt till engelska, gjorde så att man inte kan skicka in tomma listor med space och no todos när det saknas todos.
